[epgcache] Make saving EPG data optional. This allows a client receiver to passively share EPG data from another receiver over a network mount without interfering with the epg.dat file on the remote receiver.

+++ lib/dvb/epgcache.cpp
@@ -1310,9 +1310,18 @@ void eEPGCache::load()
 	}
 	else
 	{
-		unlink(EPGDATX);
-		renameResult = rename(EPGDAT, EPGDATX);
-		if (renameResult) eDebug("[eEPGCache] failed to rename %s", EPGDAT);
+		// only write to epg.dat if this is enabled in enigma2 config
+		if (!eConfigManager::getConfigBoolValue("config.epg.cachesaveenabled", false))
+		{
+			renameResult = -1;
+			eDebug("[eEPGCache::load] writing to %s is disabled in enigma2 config", EPGDAT);
+		}
+		else
+		{
+			unlink(EPGDATX);
+			renameResult = rename(EPGDAT, EPGDATX);
+			if (renameResult) eDebug("[eEPGCache] failed to rename %s", EPGDAT);
+		}
 	}
 	{
 		int size=0;
@@ -1417,6 +1426,12 @@ void eEPGCache::load()
 void eEPGCache::save()
 {
 	const char* EPGDAT = m_filename.c_str();
+	// only write to epg.dat if this is enabled in enigma2 config
+	if (!eConfigManager::getConfigBoolValue("config.epg.cachesaveenabled", false))
+	{
+		eDebug("[eEPGCache::save] writing to epg.dat is disabled in enigma2 config");
+		return;
+	}
 	if (eventData::isCacheCorrupt)
 		return;
 	// only save epg.dat if it's worth the trouble...
